function conversion
File of misc functions created to be used as exports
