+++ cmd/demote.js
@@ -2,7 +2,7 @@
 
 const { prefix, errorImg } = require('../config/config.json')
 const roleList = require('../db/roles.json')
-const rsp = require('../responses.js')
+const {rsp} = require('../functions')
 
 module.exports = {
 	name: 'demote',
