+++ cmd/mute.js
@@ -1,7 +1,7 @@
 // ========= Mutes a User
 
 const { prefix, errorImg } = require('../config/config.json')
-const rsp = require('../responses.js')
+const { rsp } = require('../functions')
 
 module.exports = {
     name: 'mute',
