+++ cmd/unmute.js
@@ -1,7 +1,7 @@
 // ========= Unmute a user
 
 const { prefix, errorImg } = require('../config/config.json')
-const rsp = require('../responses.js')
+const { rsp } = require('../functions')
 
 module.exports = {
     name: 'unmute',
