+++ discord.js
@@ -1,0 +1,7 @@
+//Set Up Discord and client to export to other files
+
+const Discord = require('discord.js')
+const client = new Discord.Client()
+
+exports.client = client
+exports.Discord = Discord
